factor out common parts of test factories
