+++ packages/benchmark/src/tests/box2d-wasm.reuse.simd.ts
@@ -8,12 +8,14 @@ import wasmSimd from 'url:../../../../node_modules/box2d-wasm/dist/umd/Box2D.sim
 
 import type { TestFactory, TestInterface } from "../types";
 
+import { box2dWasmReuseFactory } from './box2d-wasm.reuse.common';
+
 export const box2dWasmReuseSimdFactory: TestFactory = async (gravity, edgeV1, edgeV2, edgeDensity): Promise<TestInterface> => {
   if (!hasSIMD) {
     throw new Error('Cannot run SIMD test on platform which lacks SIMD support.');
   }
 
-  const { b2World, b2Vec2, b2EdgeShape, b2PolygonShape, b2_dynamicBody, b2BodyDef, destroy, getPointer, NULL } = await Box2DFactory({
+  const box2D = await Box2DFactory({
     locateFile: (url: string, scriptDirectory: string): string => {
       switch (url) {
         case 'Box2D.simd.wasm':
@@ -23,46 +25,11 @@ export const box2dWasmReuseSimdFactory: TestFactory = async (gravity, edgeV1, ed
       }
     }
   });
-  const vec0 = new b2Vec2(gravity.x, gravity.y);
-  const world = new b2World(vec0);
 
-  const bd = new b2BodyDef();
-  const ground = world.CreateBody(bd);
-
-  vec0.Set(edgeV1.x, edgeV1.y);
-  const vec1 = new b2Vec2(edgeV2.x, edgeV2.y);
-  const edgeShape = new b2EdgeShape();
-  edgeShape.SetTwoSided(vec0, vec1);
-  destroy(vec1);
-  ground.CreateFixture(edgeShape, edgeDensity);
-  destroy(edgeShape);
-
-  bd.set_type(b2_dynamicBody);
-
-  const box = new b2PolygonShape();
+  const test = box2dWasmReuseFactory(box2D, gravity, edgeV1, edgeV2, edgeDensity);
 
   return {
     name: "box2d-wasm (object re-use) (SIMD)",
-    createBoxShape(hx: number, hy: number): Box2D.b2PolygonShape {
-      box.SetAsBox(hx, hy);
-      return box;
-    },
-    createBoxBody(shape: Box2D.b2PolygonShape, x: number, y: number, density: number): void {
-      vec0.Set(x, y);
-      const body = world.CreateBody(bd);
-      body.CreateFixture(shape, density);
-    },
-    step(timeStep: number, velocityIterations: number, positionIterations: number): void {
-      world.Step(timeStep, velocityIterations, positionIterations);
-    },
-    cleanup(): void {
-      destroy(box);
-      for (let body = world.GetBodyList(); getPointer(body) !== getPointer(NULL); body = body.GetNext()) {
-        world.DestroyBody(body);
-      }
-      destroy(bd);
-      destroy(world);
-      destroy(vec0);
-    }
+    ...test
   };
 };
