+++ packages/benchmark/src/tests/box2d-wasm.standard.ts
@@ -7,8 +7,10 @@ import wasm from 'url:../../../../node_modules/box2d-wasm/dist/umd/Box2D.wasm';
 
 import type { TestFactory, TestInterface } from "../types";
 
+import { box2dWasmFactory } from './box2d-wasm.common';
+
 export const box2dWasmNoSimdFactory: TestFactory = async (gravity, edgeV1, edgeV2, edgeDensity): Promise<TestInterface> => {
-  const { b2World, b2Vec2, b2EdgeShape, b2PolygonShape, b2_dynamicBody, b2BodyDef, destroy, getCache, getPointer, NULL } = await Box2DFactory({
+  const box2D = await Box2DFactory({
     locateFile: (url: string, scriptDirectory: string): string => {
       switch (url) {
         case 'Box2D.wasm':
@@ -18,60 +20,11 @@ export const box2dWasmNoSimdFactory: TestFactory = async (gravity, edgeV1, edgeV
       }
     }
   });
-  const vec0 = new b2Vec2(gravity.x, gravity.y);
-  const world = new b2World(vec0);
-  const bd = new b2BodyDef();
-  const ground = world.CreateBody(bd);
-  destroy(bd);
-
-  const edgeShape = new b2EdgeShape();
-  {
-    const { x, y } = edgeV1;
-    vec0.Set(x, y);
-  }
-  {
-    const { x, y } = edgeV2;
-    const vec1 = new b2Vec2(x, y);
-    edgeShape.SetTwoSided(vec0, vec1);
-    destroy(vec1);
-  }
-  destroy(vec0);
-  ground.CreateFixture(edgeShape, edgeDensity);
-  destroy(edgeShape);
+  
+  const test = box2dWasmFactory(box2D, gravity, edgeV1, edgeV2, edgeDensity);
 
   return {
     name: "box2d-wasm (non-SIMD)",
-    createBoxShape(hx: number, hy: number): Box2D.b2PolygonShape {
-      const box = new b2PolygonShape();
-      box.SetAsBox(hx, hy);
-      return box;
-    },
-    createBoxBody(shape: Box2D.b2PolygonShape, x: number, y: number, density: number): void {
-      // this isn't a performant way to use Box2D.
-      // why allocate new body definitions when we could mutate and re-use one?
-      // why allocate new position vectors when we could mutate and re-use one?
-      // joelgwebber's original C benchmarks _stack-allocated_ this structure:
-      // https://github.com/joelgwebber/bench2d/blob/master/c/Bench2d.cpp#L73
-      const bd = new b2BodyDef();
-      bd.set_type(b2_dynamicBody);
-      bd.set_position(new b2Vec2(x, y));
-      const body = world.CreateBody(bd);
-      body.CreateFixture(shape, density);
-    },
-    step(timeStep: number, velocityIterations: number, positionIterations: number): void {
-      world.Step(timeStep, velocityIterations, positionIterations);
-    },
-    cleanup(): void {
-      for (let body = world.GetBodyList(); getPointer(body) !== getPointer(NULL); body = body.GetNext()) {
-        world.DestroyBody(body);
-      }
-      destroy(world);
-      // free references to everything we leaked during benchmark
-      for (const b2Class of [b2BodyDef, b2Vec2, b2PolygonShape]) {
-        for (const instance of Object.values(getCache(b2Class))) {
-          destroy(instance);
-        }
-      }
-    }
+    ...test
   };
 };
